Use a trigger to track exact number of usage counts for managed datasets

+++ webservice/src/main/java/life/catalogue/dw/jersey/ColJerseyBundle.java
@@ -60,7 +60,7 @@ public class ColJerseyBundle implements ConfiguredBundle<WsServerConfig> {
   @Subscribe
   public void datasetChanged(DatasetChanged d){
     if (d.obj!=null && d.obj.getOrigin() == DatasetOrigin.RELEASED) {
-      ccFilter.addRelease((int)d.key);
+      ccFilter.addRelease(d.key);
     }
   }
 }
